Set Logging-FATAL_MILESTONE for non-fatal CHECKs
This sets a Logging-FATAL_MILESTONE when a NotFatalUntil is provided
which will let us monitor for about-to-turn-fatal crash reports and
triage appropriately.

To support this we also refactor base/check.cc so crash keys are set in
one location and remove NotReachedDumpWithoutCrashing methods.

Bug: None
Change-Id: I7004dcf8ed08bafac6e195416957747b7567b064
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5259542
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255319}