[pip] Enforce maximum size >= minimum size for document pip
This is a speculative fix for a dcheck that shows up on CrOS.  A
window resize operation fails because the max size is smaller than
the min size.  The only path that looks like this might happen is
if there's no native window when the max size is computed -- it'll
return (0, 0) instead.  While this sometimes means "unlimited",
the resizing code interprets it as (0, 0).  Since "unlimited" is
probably not the right thing for the pip window, we just limit it
to the min size.

Bug: 1454304
Change-Id: I1584fceb54c3651698464c1e0bc82b36c1fbec65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5259159
Commit-Queue: Frank Liberato <liberato@chromium.org>
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255344}