[CCT] Fix URL view positioning
The URL view positioning goes wrong when the branding is displayed
in toolbar, since the view padding/gravity was not being restored
when the branding disappears to show title/URL. This CL ensures
the URL view padding/gravitiy are restored in that situation.
Also makes the text size larger when it is shown alone without
the title view.

Bug: 1523356, 1521366
Change-Id: I797cea0e44be3e47a070bc87fbafa98ebd5f583a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5251867
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255373}