Handle the case when EssentialSearchEnabled policy is re-disabled.
When the EssentialSearchEnabled policy is re-disabled, we want to give
the user the opportunity to explicitly choose whether they want to
accept or reject non-essential cookies and data.
To achieve this, the SOCS cookie is deleted upon EssentialSearchEnabled policy disablement. This will trigger the consent bump when using Google
search, prompting the user to accept or reject the usage of non-
essential cookies and data.

Bug: b/312542928
Change-Id: Icc4e9250a3d40fa73e161f96b7151e35b1ad3e6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5214368
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Mohammed Abdon <mohammedabdon@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255384}