[IP Protection] Tunnel HTTP requests over QUIC proxies
This CL makes proxied HTTP requests over QUIC proxies use CONNECT
tunnels instead of sending requests with the full URL to the proxy
directly. Support for configuring QUIC proxies other than for IP
Protection is being removed for now, and usage was already very low,
so this change is not expected to cause additional breakage.

This change obviates the need to fix a potential issue
identified in
https://chromium-review.googlesource.com/c/chromium/src/+/5205530/comment/678af427_eedba0ef/
where for the proxied GET requests sent to the proxy server,
HttpStreamFactory::Job::DoInitConnectionImplQuic uses
QuicSessionRequests with the destination corresponding to the
proxy server but HttpStreamFactory::Job::HasAvailableQuicSession
looks for existing QuicSessions where the destination
corresponds to the request URL.

This aligns the QUIC HTTP proxying behavior with what's
implemented for HTTP/SPDY proxying when multi-proxy chains are
in use (see:
https://chromium-review.googlesource.com/c/chromium/src/+/5085087)

Bug: 1520929
Change-Id: I1cebb46b36cecb1ef940e498563a3b052ff71b41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5241818
Commit-Queue: Andrew Williams <awillia@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255385}