Extend node count consistency check to AXTree::Unserialize()
This is a follow-up to CL:4325038 which that after deserializing a
tree update, the resulting tree matches the number of nodes
both in the renderer's a11y cache and in the serializer.
Basically, check our work. If the tree has a different number of nodes
than what is expected, that shows there's a bug somewhere.

Bug: none
Change-Id: I87b427552ceec4d90c15d49beddf6b2c437b91e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4301310
Reviewed-by: Joe Mason <joenotcharles@google.com>
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255388}