[pdf-a11y]: always load or reload a11y tree on AXMode change
In
https://crrev.com/1137918
we introduced logic to re-build (load/reload) the a11y tree when the
mode changes for PDF OCR.

This change does so regardless of whether PDF OCR is enabled or not.

Whenever a mode change occurs, RenderAccessibility is re-created and we
need to re-create the PDF accessibility tree which sets/uses state
AXPluginTreeSource/PluginAXTreeSerializer.

R=kyungjunlee@google.com

Bug: 1524128
Test: turn on VoiceOver. Navigate to a PDF. Turn off VoiceOver and wait for at least 10s before turning it back on again. Verify that VoiceOver can navigate into PDF contents.
Change-Id: I479993081ae289842e1039f04b4cdc9ac94fe02f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5259660
Reviewed-by: Kyungjun Lee <kyungjunlee@google.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255407}