[Emoji picker] Fix Polymer syntax issues
Context: https://crrev.com/c/5259339

There was a typo where [[ was not matched by ]] in a Polymer parameter, causing the expression to always evaluate as true because non-empty strings are truthy.

As a result, a separate issue went undetected. It seems that Polymer is unable to parse literals (such as true/false) in [[]] fields, only recognising method and field names.

A line has been rewritten to use one of the component fields directly, without the need to pass a boolean literal into a method.

Fixed: b:323448656
Change-Id: Ie5d7f53202e48f16cb84d69da2c33f07a8e8c4d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5260698
Reviewed-by: Chuong Ho <hdchuong@chromium.org>
Reviewed-by: John Palmer <jopalmer@chromium.org>
Commit-Queue: Luke Fisk-Lennon <lukefl@google.com>
Cr-Commit-Position: refs/heads/main@{#1255440}