[Office upload] Dump state when the uploaded url has an unexpected host
The uploaded (to Drive) url is always expected to have a
"docs.google.com" host. But we have seen from the metrics that this is
not always the case as there are kUnexpectedAlternateUrlHost values
logged for the OfficeFilesUploadResult metric.

Dump the state with unique crash keys to determine what other hosts are
appearing. Leave a TODO(b/323452926) to remove this dumping once the
root case is found.

The query in go/crash would be:
```
EXISTS (SELECT 1 FROM UNNEST(productdata) WHERE
key='OfficeUpload-UnexpectedHost') AND product_name="Chrome_ChromeOS"
```

Bug: b/322909593
Change-Id: Iea2f9d3b05470b93418106a3db4aab0494277c91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5260279
Commit-Queue: Cassy Chun-Crogan <cassycc@google.com>
Reviewed-by: Alexander Bolodurin <alexbn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255441}