Adjust CSSSelectorList tracing
We're improving the clang plugin for blink gc so that it alerts on
untraced array of traceable objects. Due to other technical limitations
of the plugin, it is not able to deduce that first_selector_ is traced
from the current implementation of the Trace method. Specifically, it
needs to see that the field is accessed (the check can't peer into the
First() getter method). This CL adjusts the Trace method so that the
plugin can correctly "understand" it.

Bug: chromium:1480476
Change-Id: I2c035003483c6c4ffff22040498518fab7476d85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5253450
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255465}