[Autofill] Replace unique_renderer with renderer_id 7/8
This CL replaces `unique_renderer` with `renderer_id` in the listed
files.

This CL was generated by
$ sed -i 's/unique_renderer_id/renderer_id/g' <affected-files>
and then reverting all changes in string literals.

That is, this CL only affects the C++ code; the JavaScript and
TypeScript code as well as JSON objects are left unchanged.

Bug: 1522715
Change-Id: I3ae4136345dc9891a90e09e5ff12002e9f3fc6fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5255477
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Christoph Schwering <schwering@google.com>
Reviewed-by: Maria Kazinova <kazinova@google.com>
Cr-Commit-Position: refs/heads/main@{#1255487}