[InterceptBubble] Adapt identity pill effect for each intercept type with a bubble
The changes were already existing for the ChromeSigninBubble, the goal
of this fix is to extend them to the Profile Separation bubble and the
Profile Switch bubble.

- While the bubbles are shown, do not allow the profile menu to be
displayed on top of it. The button action will be disabled.
- The identity pill extends and show a text related to the bubble
type being displayed, while the bubble is disaplyed and the button
action is disabled.

Bug: b:319410070
Change-Id: I6a57a37280eddde82cfebdda185114c470f9667c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5250584
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1255491}