Generate values for synthesized nodes during FinalizeAfterImport.
During profile finalization, we now generate values for the tree's
synthesized nodes. Note that these values are generated regardless of
which kind of update was made on the tree values, whether it was tree
completion or gap filling of nodes.

The values for synthesized nodes are generated using i18n formatting
expressions.

Bug: 1516912
Change-Id: I74e6281576c75328b699e9090751d986591c7707
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5250238
Reviewed-by: Dominic Battre <battre@chromium.org>
Commit-Queue: Norge Vizcay <vizcay@google.com>
Cr-Commit-Position: refs/heads/main@{#1255531}