[UNO] Show chrome sign in bubble if chrome isn't signed in
This CL changes the Chrome sign in bubble to always show it if:
- the user isn't signed to chrome
- and sign in is happening through the web

This changes from the previous behavior that the bubble will also be
shown on reauth and for secondary accounts.

This CL also changes the priority of recording
`ShouldShowChromeSigninBubbleWithReason`. It will be easier to
understand the data if we first exclude sign in that aren't happening
through the web. This will help exclude users that sign in through
chrome from the bucket `kShouldNotShowAlreadySignedIn`.

Bug: b/320426816
Change-Id: I83d93bf9369f1ed627d90b0b48e65ad0ad0daf92
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5224249
Commit-Queue: Monica Basta <msalama@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255554}