immersive fullscreen: make toolbar style optional
The goal of this CL is to reduce the number of UpdateToolbarVisibility()
calls. We do this in a couple of ways. First the last_used_style_ is now
optional. Immersive fullscreen is enabled before we know what style
should be used. We previously had a default value but this causes churn
when eventually we do get the desired style. This is a factor in
1519600. The optional toolbar style allows us to ignore style updates
when the value is not set.

By reducing the number of calls to UpdateToolbarVisibility() we also
reduce the number of calls to -forceVisibilityRefresh which also
contributes to 1519600 and is the cause of 1519591.

Bug: 1519600, 1519591
Change-Id: Ia0b51241bac0dd850560e0ead72241e7d63bd4e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5259100
Commit-Queue: Tom Burgin <bur@chromium.org>
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255597}