[iOS fuzzer] Fixed crashes in chrome_app_startup_parameters fuzzer
Fixed two issues with crashing in the three fuzzers for
chrome_app_startup_parameters. An AtExitManager was added into the
fuzzer's environment, which is needed as a callback for CHECKs that
naturally fail in the fuzzer's limited environment. Also, we
instantiate the i18n library since it is needed by GURL inside the
fuzzed function with i18n::InitializeICU().

Bug: 1523661
Change-Id: I8a4e6bf2b3410c65b119fad006c3ac24fb34145e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5253738
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Nicolas MacBeth <nicolasmacbeth@google.com>
Cr-Commit-Position: refs/heads/main@{#1255612}