Fix Omnibox search suggestions when third-party cookies are blocked
The issue is caused by the fact that StartSuggestionsRequest and
StartDeletionRequest were not setting the request's SiteForCookies
to https://google.com.

As a result, the requests are considered 3P by the network process and
if 3P cookie blocking is enabled, the HTTP requests spawned by these
IPCs will not have cookies attached.

Bug: 1521141
Change-Id: I09dd5abf7065f690f3ab101dcc8cf17c3a959a9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5259121
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Dylan Cutler <dylancutler@google.com>
Cr-Commit-Position: refs/heads/main@{#1255636}