Move ResourceUsageTabHelper to performance controls folder
Moving and re-naming the ResourceUsageTabHelper helps de-couple the tab
helper from the UserPerformanceTuningManager. This also lays the ground
work for having the tab helper work with the resource attribution API
instead of relying on the UserPerformanceTuningManager to populate tab
resource usage.

This CL won't be submitted until https://crrev.com/c/5214772 is ready
to submit to prevent the dependency of the UserPerformanceTuningManger
on c/b/ui/performance_controls/

Bug: 1506236
Change-Id: I70a2fa7624fa0117a3d6de242ecd9b13c7094dd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5211415
Reviewed-by: Joe Mason <joenotcharles@google.com>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Commit-Queue: Steven Luong <stluong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255651}