Remove sync suppression logic while local UPM migration is pending
Behavior changes are behind flag.
Partial revert of crrev.com/c/5088712. The suppression logic will
have to be implemented differently, remove for now. After this CL it is
possible to enable password sync while the migration is pending, but in
*transport* mode, without any upload of existing data.
This setup of overriding the SyncMode to kTransportOnly in the
controller can potentially be reused on Desktop in the future, if we
decide to always run passwords in transport mode ahead of fully
deprecating sync-the-feature. So gate it behind a new
kEnablePasswordsAccountStorageForSyncingUsers toggle. This toggle
will also be consumed by IsOptedInForAccountStorage() in future CLs.

Unrelated to that, also add a comment next to the usage of the
corresponding flag for non-syncing users.

Bug: 1509058
Change-Id: I70fe3497e46bb0f03611fa96844bc71eb6f3a300
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5262157
Auto-Submit: Victor Vianna <victorvianna@google.com>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Commit-Queue: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255656}