Keep upload request as protos till client
Make it so that records to be uploaded remain protobufs until
`EncryptedReportingClient` level and convert to JSON only there.
To support that, traffic UMA is also moved to the same level.

The ultimate goal is to allow records to be cached - the cache will
become part of `State` that is currently located in
`EncryptedReportingJobConfiguration` and based on JSON.

In the process I moved protos->JSON request conversion from UI thread to thread pool, and then had to separate detection of mananged/unmanaged device, which is the only part of it that must be run on UI. It has been converted into a parameter of JSON builders, set up once per conversion.

Tested manually on VM and by unit and browser tests.
Also tested by `git cl try -B luci.chromium.try -b win-celab-try-rel`

Bug: b:314195198
Change-Id: I57c0dca9d15a553db76cdef12de79c5f4532c59a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5253304
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Josh Hilke <jrhilke@google.com>
Cr-Commit-Position: refs/heads/main@{#1255689}