Fixed const-correctness for BaseRenderingContext2D::GetPaintCanvas
We have two GetPaintCanvas getters, one const and one non-const.
Because this method is virtual, a const_cast scheme is used to avoid
having all child classes define that method twice. The scheme was done
the wrong way though: the const method was calling the non-const
virtual method. This means that if any of the implementations of the
non-const overload was modifying the object, the const overload would
also modify the object. It's safer to have the non-const overload call
the const one.

Bug: 1484741
Change-Id: If294fbdf7c3227d19439a8b4ed2783850a10b01a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5249732
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255738}