Preserve recorder when replacing the Canvas2DLayerBridge
In the follow-up CL (https://crrev.com/c/5258875), the recording will
be split to store the content of layers into a separate recording.
That side recording cannot be flushed until the layers are closed.
This means that when replacing the Canvas2DLayerBridge, we cannot
rasterize that part of the recording into the raster that is
transferred to the new bridge. We instead have to preserve the layer's
recording itself, which can be done by transferring the whole recorder.

Bug: 1484741
Change-Id: I9a6dff2226a238493e5f552d81e375b89313429b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5258874
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255741}