[bookmarks] Adjust save flow dimens to match specs
When looking into the chevron overlap, I noticed there was more space
that could be left on the end of the view to fit the folder name. Then
I checked the specs and noticed that there were a couple of dimens that
didn't match the specs which would allow for more space.

Bug: 1522770
Change-Id: Ia6eaeca341ce768c32b1ac322e9a2962d4a74673
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5254197
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: Brandon Wylie <wylieb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255745}