[live-preview] Fix media preview scroll view crash
This `SetViewportRoundedCornerRadius(..)` function should only be called if this `kUiCompositorScrollWithLayers` feature is enabled, otherwise it will crash. This feature appears to be only enabled on Mac for now.

Removed the usage of `SetViewportRoundedCornerRadius` for now. Added a TODO to figure out a better solution.

Bug: b:321853311,b:323568011
Change-Id: Ibb6f21d92879333a6397062d44997f0eaa063352
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5263178
Auto-Submit: Ahmed Moussa <ahmedmoussa@google.com>
Reviewed-by: Ahmed Moussa <ahmedmoussa@google.com>
Commit-Queue: Ahmed Moussa <ahmedmoussa@google.com>
Reviewed-by: Bryant Chandler <bryantchandler@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255765}