iwa: Add IsolatedWebAppBuilder
This adds an IsolatedWebAppBuilder class, along with supporting
ManifestBuilder, and Scoped[Bundled|Proxy]IsolatedWebApp classes. These
make it easier to specify the contents of an IWA in tests for both
proxy and bundled use cases.

This CL doesn't add PageState mocking for unit tests. That will be
added in a followup CL, and will enable this builder to be used in
unit tests as well.

Bug: b/230033381
Change-Id: I7e3ba7ca001b9a332d135ab070555c787dce2719
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5252172
Reviewed-by: Zelin Liu <zelin@chromium.org>
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: Oleksandr Peletskyi <peletskyi@google.com>
Cr-Commit-Position: refs/heads/main@{#1255781}