desk_profiles: Do not restore to the previous desk for new browser
With Desk Profiles feature, chromeos can now associate a lacros profile
with a virtual desk. When launching a new browser within the desk, a
browser for that profile will be launched. With session restore, it
restores the new browser window to its previous desk.

This CL proposes that when lacros and Desk Profiles are both enabled,
the new browser window should *not* be restored to its previous desk.

Quick demo can be found at - b/319730718.

Bug: b/319730718
Test: manual
Change-Id: I20360313714ac152d8c681f3d1d47e4b88e3f383
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5249158
Commit-Queue: Yongshun Liu <yongshun@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255786}