Fix Hub interactions with OmniboxMatchToolbarAndStatusBarColor.
When this feature was enabled, the Hub did not emit alpha change
events at all. This would cause the status bar color to jump or not
update. To fix this new observer plumbing was added, to get the
observer up out of the TopToolbarCoordinator and into Hub animations.

However this still was not sufficient. As this exposed that the start
toolbar was being shown (and emitting alpha values) for certain
feature configurations. This is UI surface is now correctly hidden in
ToolbarManager#updateForLayout when Hub is enabled.

Lastly, the incognito state the ToolbarColorObserverManager read from
IncognitoStateProvider which proxied from the TabModelSelector would
update too quickly when new tabs were created and animated from the
Hub. To fix this an incognito supplier override is now plumbed through
to use when available.

Change-Id: I083230662f7aba9985f01e9b4e749b370f26c17c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5251264
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Sky Malice <skym@chromium.org>
Auto-Submit: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255811}