Correctly removeObserver on consecutive showScrim calls.
Previous CL https://crrev.com/c/5234853 added a model observer in
showScrim, and removed it in hideScrim. However when two showScrim
calls are made with no hideScrim, the model is replaced, but no
removeObserver is ever issued.

To fix this, we move around the setAlphaInternal inside showScrim so
that it cleans up the observer before we switch mModel.

Bug: 1523529
Change-Id: Ifb50c6e9671770b398c0b3be41d8e6f6ef2d3caf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5263062
Commit-Queue: Sky Malice <skym@chromium.org>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255813}