[automation] Remove unused permission `matches` from manifest
As part of the cleanup now that the automation API is fully private
(go/chrome-automation-partial-deprecation), remove a manifest
permission that is not used.

The permission will be converted to a single boolean in a follow-up
change.

AX-Relnotes: n/a.
Bug: b/265481580
Test: All automation tests
Change-Id: I04573735cf76172990fd6b6d9281d5620b5eb977
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5226180
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Auto-Submit: Anastasia Helfinstein <anastasi@google.com>
Cr-Commit-Position: refs/heads/main@{#1255829}