desk_profiles: Fix reverse tabbing in the desk bar
This CL fixes reverse tabbing in the desk bar. This would have not been
an issue to begin with if it hadn't been for the trickery that is done
with hiding the desk action buttons as focus shifts. This CL extends the
hack that already involved the desk preview view to now also include the
profile button.

Also made changes to the test suite so that the DeskBarTest suite runs
with desk profiles both disabled and enabled.

BUG=b:319723846

Change-Id: I462f58123f2a3dbb5babc2ac02c9135c4fb5831b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5263242
Commit-Queue: Daniel Andersson <dandersson@chromium.org>
Reviewed-by: Yongshun Liu <yongshun@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255855}