IDB: move more of IndexedDBFactory into IndexedDBBucketContext
This moves the rest the code that sets up and tears down IDB
implementation classes like IndexedDBBackingStore from
IndexedDBFactory to IndexedDBBucketContext. IOW, this makes
IndexedDBBucketContext the mojo endpoint for `window.indexedDB`
API calls like `open()` and `deleteDatabase()`.

This is necessary because IndexedDBFactory shouldn't be reaching
into bucket sequence classes, and also because the renderer
should talk directly to the bucket sequence thread instead of
making an additional stop on the factory thread. What remains in
IndexedDBFactory should be folded into IndexedDBContextImpl.

Notably, and perhaps treacherously: this affects the lifetime of
the IndexedDBBucketContext. Formerly it was created when an api
method like `open()` or `deleteDatabase()` was invoked, and
discarded if it was not in use for over 2 seconds. With this
change, it is created when the first receiver is bound and
destroyed when the last receiver is unbound. The backing store
is still destroyed when not in use for over two seconds (via
`IndexedDBBucketContext::ResetBackingStore()`).

Low-Coverage-Reason: COVERAGE_UNDERREPORTED inspector tests aren't being reported
Bug: 1474996

Change-Id: Ibf5e7546e2bc21d37593845f8da4faa4da0607dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5173704
Reviewed-by: Emily Stark <estark@chromium.org>
Reviewed-by: Austin Sullivan <asully@chromium.org>
Auto-Submit: Evan Stade <estade@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255961}