Add Managed Client Certificate Tab to Connectors Internals Page
This new tab will help debug whether the provisioning of a managed
identity was successful or not, along with other properties. As with
the existing Device Trust tab, this UI is not meant for actual
end-users.

https://screenshot.googleplex.com/Ae69ZZ8NEnoSLpH

This CL also fixes and issue with the way tabs were getting a reference
to the pageHandler, making one tab overwrite the Mojom remote of the
other (effectively failing the request).

Bug: b:322838752
Change-Id: I82f3a6d80c8684827d71f9994956e9e47bb8565b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5259183
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Hamda Mare <hmare@google.com>
Commit-Queue: Sébastien Lalancette <seblalancette@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1256014}